requirements.txt
# === Core ===
torch==2.4.*
torchvision==0.19.*          # optional but useful
torchaudio==2.4.*

# === Reinforcement Learning stack ===
torchrl==0.4.*               # matches Torch 2.4
tensordict==0.4.*            # TorchRL dependency (explicitly pin to same minor)
gymnasium==1.0.*             # modern Gym API
minari==0.4.*                # for offline RL / BC datasets
dm_env==1.6                  # used by some TorchRL wrappers

# === Hydra & HPO ===
hydra-core==1.3.*
omegaconf==2.3.*
hydra-optuna-sweeper==1.2.0
optuna==2.10.1
SQLAlchemy<2.0


# === Logging & Experiment Tracking ===
wandb==0.17.*                # or latest stable you use
tqdm==4.*
rich==13.*

# === Data / Utils ===
numpy>=1.24
pandas>=2.0
scipy>=1.11
pyyaml>=6.0
matplotlib>=3.8              # optional for debug viz

# === Environment helpers ===
# (fetch environments and similar robotics)
mujoco==3.1.3                # needed by Gymnasium robotics
pygame==2.5.2                # optional; for visual envs
imageio==2.34
imageio-ffmpeg>=0.4
opencv-python-headless>=4.10 # for rendering in headless docker

# === Docker safety / reproducibility ===
# versions below ensure deterministic builds
protobuf<5.0
